refactor: mock-data 폴더로 모의 데이터 경로 이동
