+++ src/main.js
@@ -1,5 +1,5 @@
 import { $ } from "./utils/domHelpers.js";
-import { restaurants } from "./constants/restaurantData.js";
+import { restaurants } from "./mock-data/restaurantData.js";
 import modalHandler from "./eventHandler/modal.js";
 import renderMainPage from "./views/mainPage/mainPage.js";
