+++ src/views/mainPage/components/buttonContainer.js
@@ -1,4 +1,4 @@
-import { restaurants } from "../../../constants/restaurantData.js";
+import { restaurants } from "../../../mock-data/restaurantData.js";
 import restaurantList from "./restaurantList.js";
 import { $ } from "../../../utils/domHelpers.js";
 import button from "../../../components/@common/button.js";
